added existence check to train_cpvae when trying to load decision tree

+++ sounds_deep/contrib/experiments/train_cpvae.py
@@ -305,10 +305,12 @@ with tf.Session(config=config) as session:
                 saver.save(session,
                            os.path.join(args.output_dir, 'model_params'))
 
-                with open(
-                        os.path.join(args.output_dir, 'decision_tree.pkl'),
-                        'wb') as dt_file:
-                    pickle.dump(model._decision_tree, dt_file)
+                decision_tree_path = os.path.join(args.output_dir, 'decision_tree.pkl')
+                if os.path.exists(decision_tree_path):
+                    with open(
+                            decision_tree_path
+                            'wb') as dt_file:
+                        pickle.dump(model._decision_tree, dt_file)
                 exit_fn.best_class_rate = class_rate
             # not currently doing early stopping
             return False
